Workaround to Travis-CI Wine/Mingw build hanging occasionally
Summary:
As you may recall, I investigated T44 and determined that it's a bug with Wine on Ubuntu 14.04.  As such, occasionally and randomly, mingw builds fail on travis with no output for 10 mins.

On my Ubuntu 14.04 system I can also reproduce the hang (very occasionally) and when I do, indeed bitcoint_test.exe takes forever to run but it does complete.

So, as a workaround, I added travis vars and a small modification to test_bitcoin to print output to the console.  Note: in order to do this I had to replace the default automake test logger with something that also spits to console.

This has been testedon travis and only in Windows builds on travis do we get the output.  On linux and other builds everything works as before.

Test Plan:
push a commit to GitHub on a travis-enabled repository and see if the windows builds error out with the '10 mins no output' bug.  They shouldn't.

Here is a sample Travis build that worked perfectly.  Note the Windows builds had console output to keep them alive:

https://travis-ci.org/cculianu/bitcoin-abc/builds/268807063

Reviewers: freetrader, deadalnix, #bitcoin_abc!

Differential Revision: https://reviews.bitcoinabc.org/D497